+++ internal/model/request/relation.go
@@ -2,12 +2,12 @@ package request
 
 type RelationActionReq struct {
 	Token      string `form:"token" json:"token" binding:"required"`
-	ToUserID   int64  `form:"to_user_id" json:"to_user_id" binding:"required"`
+	ToUserID   uint   `form:"to_user_id" json:"to_user_id" binding:"required"`
 	ActionType int32  `form:"action_type" json:"action_type" binding:"required"`
 }
 
 type FollowListReq struct {
-	UserID int64  `form:"user_id" json:"user_id" binding:"required"`
+	UserID uint   `form:"user_id" json:"user_id" binding:"required"`
 	Token  string `form:"token" json:"token" binding:"required"`
 }
